move some utility functions to the util file.
